+++ modules-services2/service-templates-core/src/main/java/org/openprovenance/prov/service/core/TemplateService.java
@@ -778,7 +778,7 @@ public class TemplateService {
         }
     }
 
-
-
-
+    public TemplateQuery getQueryTemplate() {
+        return queryTemplate;
+    }
 }