+++ modules-tools/prov-template-archetype/src/test/resources/sql/fs-tables.sql
@@ -1,0 +1,2 @@
+
+SELECT * FROM count_rows_in_tables(ARRAY['activity', 'agent', 'file_transforming', 'file_init' ]);