+++ modules-tools/prov-template-archetype/src/test/resources/sql/test2.sql
@@ -1,0 +1,3 @@
+select *
+from activity
+WHERE created_at IS NOT NULL ORDER BY created_at DESC LIMIT 1;